+++ tests/testthat/test-saem-translate.R
@@ -64,8 +64,6 @@ test_that("Standard theo linCmt()", {
     })
   }
 
-  skip_if_not(rxode2::.linCmtSensB())
-
   f <- .nlmixr(one.cmt)
 
   expect_equal(f$saemModel0,